Position scrollbar shows current position and can be used for seeking
Adjustment::get_value returns a double. This value is interpreted as
seconds. For consistency MediaPlayer and GtkmmMediaPlayer do also use
double values holding seconds on its user interface.

Player internally 64 bit integers are used with different timebases. This
is how FFmpeg is handling time stamps.

+++ gui/GtkmmMediaPlayer.hpp
@@ -18,8 +18,8 @@ class GtkmmMediaPlayer : public MediaPlayer
 
 public:
     sigc::signal<void, std::string> notificationFileName;
-    sigc::signal<void, int> notificationDuration;
-    sigc::signal<void, int> notificationCurrentTime;
+    sigc::signal<void, double> notificationDuration;
+    sigc::signal<void, double> notificationCurrentTime;
 
     GtkmmMediaPlayer(boost::shared_ptr<PlayList> playList);
